Allow the shunting of in-progress migrations
The "native" flag in the provider interface seems to be an issue. As we
define our mappings (migrations or container mappings), they are
intended to refer to a specific storage location. In this context, it
doesn't make sense to make the notion of whether we're sharding data in
S3 to be overridden per request.