Add context in status to identify all bucket migrations
Status records for migrations configured with aws_bucket "/*" replace
their container and aws_bucket key values with the name of the remote
bucket.

It's useful when reasoning about the migration status records to
differentiate between the status record of a migration for a single
container or the multiple status records generated for migrations of all
the containers in the account.

Add a 'all_buckets': True to the status records for migrations
configured with 'aws_bucket': '/*'