Do not fail mismatched ETags (likely, due to *LO).
When we check if an object has been migrated, large objects present a
unique challenge. From the remote cluster, the large object will appear
to have a different ETag, as it will be the hash of hashes of individual
segments.

On the destination cluster, with InternalClient, we will receive the
ETag of the manifest itself. Because of this mismatch, we can't rely on
ETags to detect subtle corruption (where the source and destination
ETags differ, but the Last-Modified date is the same).

This patch does two things: logs a message for exceptional cases and
ignores cases of large objects (while doing its best to compare the
actual manifests for SLOs).