Do not double release a connection semaphore.
If we double release the connection semaphore, the semaphore count goes
to 2, resulting in the connection being available to be checked out
twice. This can lead to data corruption, as multiple co-routines may be
using the same connection.

The patch adds some tests and a check for the semaphore balance before
calling release to try to detect such bugs.