First shot at a minimal cloud-connector daemon
The cloud-connector is designed to run in a container in a public cloud
whose "local" object storage uses the S3 API.  Its configuration is
stored in that "local" object storage and the daemon's container
requires enough configuration via environment variables to be able to
find its main config file, and a copy of the cloud-sync JSON config
file.

The idea is to provide a Swift or S3 API in the public cloud which has
"seamless" access to data in the on-prem Swift cluster.  The primary
use-case is bursting workloads to the public cloud which normally run in
private cloud, using the on-prem Swift cluster.

Further work will expand cloud-connector's capabilities and unlock more
use cases.  It will grow into its name.

This first version only supports a rudimentary GET and HEAD path for
objects.

But the infrastructure to be able to write integration tests against it
(and interact with it in the container via a bash shell) is there.

Added a tempauth Unicode user, account and password; sadly had to patch
tempauth AND python-swiftclient to handle the Unicode biz.

available in our tempauth config as well as swiftclient... this way we
can test cloud-connect with lots of Unicode business.  Woohoo!

TODO: the actual tests.