Handle Swift versioning in migrations.
When migrating containers, we should consider Swift versioning and
handle it correctly. There are two types: history and versions. One
behaves like a historical view of the container/objects and the other is
more akin to a stack (like S3 versioning), where objects are popped off
the stack as versions are deleted.