Integration test helpers
...for getting swiftclient.client.Connection() instances pointed at any
account, and going through the normal proxy, a noauth/noshunt proxy, or
the cloud-connector.

The good stuff is:
    @classmethod
    def conn_for_acct(klass, acct):
        ...

    @classmethod
    def conn_for_acct_noshunt(klass, acct):
        ...

    @classmethod
    def conn_for_acct_cc(klass, acct):
        ...

One conn per all accounts discovered during setup are stored in
    conn_by_acct = {}  # utf8_acct => Connection()
    conn_by_acct_noshunt = {}  # as above, but bypasses any shunt
    conn_by_acct_cc = {}  # as above, but through cloud-connector

So tearing down all objs in all containers in all accounts is as easy
as:
        for conn in klass.conn_by_acct_noshunt.values():
            clear_swift_account(conn)

A lot of the old stuff was left alone to minimize test disruption,
e.g.
        klass.swift_src = klass.conn_for_acct('AUTH_test')
        klass.swift_dst = klass.conn_for_acct(
            u"AUTH_\u062aacct2".encode('utf8'))