Propagate DELETE container requests in migrations.
As we cannot set the X-Timestamp header when creating a container during
migrations, we have to make sure to propagate removal of containers to
the source cluster. As we don't do this currently, the containers will
be re-created on a subsequent migration run.

The commit also mutates the associated test to examine
the content_location information for the listings. This is necessary, as
after adding the support for migrations in the shunt we can no longer
simply examine the listing names.