Migrations: send container metadata to the source.
During a migration, a container may not have been yet created on the
destination cluster. At this time, if a POST request is made, we should
propagate it to the source cluster. Returning a not found error in this
case would be confusing to the caller, as they would be able to GET and
HEAD the container, just not issue a POST.