ProviderResponse.to_wsgi returns a status string.
The WSGI contract is that the status passed to start_response() should
be a string consisting of an HTTP status and HTTP reason. Previously,
the call to to_wsgi() method in ProviderResponse would return an
integer and we had some hooks to convert the status to a string in some
places.

The goal of this patch is to standardize the behavior where
ProviderResponse.to_wsgi will always return a valid status string and we
do not have to worry about converting it. It also simplifies the logic
when constructing the responses, as we can always create the
ProviderResponse object with an integer status. The Swift
swob.RESPONSE_REASONS is used for the lookup of HTTP reason.