Remove an extra HEAD request when shunting LIST.
When we shunt a request a GET request on the container for a migration,
we currently make a HEAD request because of a deficiency in the
list_objects API. Unlike the other provider API methods, it returns a
tuple of status and body. Because headers are missing from the response,
we make a HEAD request.

The patch changes the API to return a ProviderResponse (like the other
methods) and refactors the listing handling to take advantage of the
additional information. Majority of the changes are related to fixing
the tests for the API change.

The shunt listing logic is also simplified somewhat.