Fix the shunt for migrations with all containers.
When a migration is configured for all containers, we need to update the
remote container field in the sync_profile (aws_bucket), as otherwise
the remote container would be "/*", which is not a valid container. The
behavior manifests itself as the shunt seemingly not working, as all
requests will return 412 (Bad URL error in the Swift proxy container
controller).