Submit SLO ETag as a header on restore/migrations.
We had a prior bug where SLO segments included the ETag in the WSGI
environment, which doesn't get set on the request. We should submit it
as a header instead.

This patch also requires Swift to be at least version 2.13 for cloud
sync, as we cannot validate the SLO ETags otherwise on uploads.