Create container on PUT object for migration
The shunt will now create the container when PUT object attempted on a
container which has not yet been migrated. Currently, we end up doing an
uncoditional HEAD on the container. This is not ideal. If we can figure
out a way to keep the input stream open, we can remove the HEAD request
in the common case (when the container exists).

Co-Authored-By: Timur Alperovich <timuralp@swiftstack.com>