Use the proper ETag for SLO and content for DLO.
Our SLO/DLO uploads broke and we missed the 422 errors when attempting
to upload the manifests. There are two issues. For the SLO case, we
should be uploading the ETag of the *JSON content* rather than the ETag
returned by the origin Swift cluster (as there is no SLO middleware to
validate it).

For the DLO case, we have to GET the object with multipart-manifest=get
so that we retrieve the actual manifest blob. This is critical because a
DLO allows the manifest object to be non-zero size.

The patch fixes the large object tests that stopped working after the
migration shunt was put in place. There is also a small logger issue,
where we were no longer logging the actual requested objected on certain
failures.