Do not make a HEAD request on every object PUT.
When migrating and the target container does not exist, we have to guard
against returning a 404. This may be confusing to the client, as the
container will appear to exist, but is only available in the cluster
that is the source for the migration.

Previously, we worked around it by making an unconditional HEAD request.
A better approach would be to attempt to PUT the object and if that
fails, create the container.