Add S3 API authen/authz for cloud-connector
We decided to only support the S3 API in the first pass of
cloud-connector.  So the auth had to be rewritten and swift3
incorporated.

cloud-connector now requires a DB mapping S3 API key IDs to secret key
values and Swift Account names.

There still aren't any integration tests because the cloud-connector's
controller is still a dumb stub that just returns NotImplemented() for
all verbs.

[#156747601]