Fix a Unicode bug; improve error logging
Also fix a place where [''] was returned, which isn't an iterator.

And another spot where iter('a string') returned a bunch of chars vs,
just iter(['a string']).

This stuff fell out of a bunch of other speculative changes and while I
don't remember or have the triggering condition that _hit_ this unicode,
bug, this new code is definitely robust to having a Unicode instance
already passed in, where the old code was not.

The improved error logging in the S3 sync provider was helpful to me
when using that provider in the cloud-connector code when I (as a human)
couldn't tell in any longs wth was going on when the boto thing was
failing.