Delete migrated objects if source is removed.
When source objects are removed and there were no changes to the
destination objects, the migrator should remove the destination objects.
This allows us to handle properly a case where clients continue to
mutate the state of the source object store before switching over to
using the destination store (after some number of objects have been
migrated).

The approach introduces a Swift SYSMETA header that is used to tag all
objects on PUT into the destination. When an object's metadata is
updated or a PUT is issued, this header is overwritten. This allows us
to detect changes and delete objects only if they have not been by
mutated on the destination cluster.