Add provider put_object() methods.
The cloud-connector will need to be able to just PUT an object, to
either S3 or Swift, sanely (i.e. accepting an iterable of body, not
requiring a known length).

Also beefs up some of the "wrapper" helper classes.

All the new and changed code has 100% line and branch unit test
coverage, AFACT.