Add cloud-connector obj GET/HEAD support
When a S3 API object GET/HEAD request comes into cloud-connector, it
first looks in the local S3 API object store, and if that "misses", it
looks in the configured on-prem Swift object store (using the S3 API).

Also made create_provider able to plumb a logger into the provider.
Cloud-connector uses that to ensure that provider logging shows up and
is appropriately prefixed.

[#156747726]