Migrator: remove deleted source containers.
When a container is removed from the source, we should remove it from
the destination cluster during migrations. In this case, however, we
need to be cognizant of multiple scenarios:
1. container copied to the destination with no changes
2. container copied to the destination and had its metadata updated
3. container copied to the destination and had objects added

Only in the first case should we be removing the container from the
destination. In the second case, we should remove all objects, but
preserve the container and its metadata. In the third case, we should
remove only objects that were present in the source container, but
preserve any new objects (once that were PUT on the destination).

We rely on the shunt and system metadata to keep track of the "state" of
the container. When a container is copied, it is in the MIGRATING state.
Once the container had a POST, it is in the MODIFIED state. A container
in the MODIFIED state can never be removed. However, to avoid scanning
for objects in the container, after we remove any objects we should, the
container is marked as SOURCE DELETED. If the source container is
recreated, the container can transition back to the MODIFIED state (but
never MIGRATING, as that would mean the container is again eligible for
removal).