Do not block when listing containers.
When we call InternalClient.iter_objects(), we have to hold the internal
client instance checked out of the pool until we are done with iterating
through the response. The patch adds our own iterator that can check out
the internal client instance from the pool. This allows us to alternate
downloading objects with listings, even when there is only one
connection.