Tooling to build a cloud-connector Docker container
We can now build a container for cloud-connector.  Furthermore,
cloud-connector only runs in its own container, even during development
and integration test runs.

Unified how CI and developers run tests, including how the involved
containers are built and started.  The cloud-connector for tests is now
in its own container.

The use of docker-compose is mostly deprecated at this point, and when
switching to this code tree and beyond, you'll want to stop and "rm" the
docker-compose-related container, if you have one.

You can override the default service host port mappings by creating a
file named port-mapping.env` in the root directory of this code tree
with other values in this format (it is sorced by a shell):
  HOST_S3_PORT=...
  HOST_SWIFT_PORT=...
  HOST_CLOUD_CONNECTOR_PORT=...

[#157237645]