Set the level of some debug messages to INFO.
There are a number of messages that we currently log at DEBUG that are
both infrequent and helpful in understanding how the system works. We
should log these at the INFO level. Specifically, creating containers
and processing stats (which migration is being processed and how long it
took). The patch also includes the account in the "Container not found"
errors, so that it's easier to check which account is in question.