cloud-connector: Add obj metadata update support
Via S3 API, what Swift calls an object POST is spelled as a PUT copy request
with copy source same as path and X-Amz-Metadata-Directive header set to
"REPLACE".

This now works, with semantics that the metadata update is applied to
local store if possible, but if that fails, the post is applied to the
remote store.

The providers grew a new method, "post_object" which does about what you
expect.

NOTE: when using S3 API to "post" to an object through swift3
middleware, new metadata is apparently merged with old metadata; I
assume this is an artifact of the Swift copy middleware being involved.

[#156747759]