Resolves a deadlock when migrating DLOs.
When migrating a container with numerous DLOs, we can encounter a
deadlock if all of the workers end up enumerating the DLO container and
none of the workers are draining the upload tasks. The patch reorganizes
the work such that all referenced (DLO) containers are submitted into a
new queue. The enumerating thread waits for all the work to be complete
for a set of containers in the queue before processing any additional
containers that may have been found (allowing us to handle nested DLOs).