Do not block on putting segments into the queue.
When the migrator is copying SLOs and their segments, there may be a
situation when there are more segments than our queue can accommodate.
If all workers pick up an SLO, they can saturate the queue such that no
progress can be made. In that case we should process the segments within
that worker (and not put the work into the queue).