Fix the flaky propagate_delete test.
When we moved from using X-Timestamp to Last-Modified we introduced a
possibility of creating objects with a timestamp in the future. This
happens because Last-Modified date, unlike X-Timestamp, is computed as
math.ceil(file_x_ts). We are forced to rely on the Last-Modified date
because X-Timestamp does not include the metadata offset.

The test now attempts to detect how much time drift we introduced and
sleeps for that amount of time to guarantee that the request will happen
at or after the object creation time.