cloud-connector: automatically reload config files
Automatically reload updated sync or S3 DB config files in S3-api object
store.  We don't reload the main config file because there are lots of
settings in there that require a restart anyway.

Each of the c-c application class and the c-c
authentication/authorization classes inherit from ConfigReloaderMixin
and define a CHECK_PERIOD property that defines the minimum time between
checks against the S3-api object store holding the config files.  The
ETag of fetched configs is cached and used with IfNoneMatch to put the
onus of detecting change on the S3-API object store.

[#156747715]