Remove migrations not handled by the node.
When we prune migrations after handling all containers for an account.
Previously, we would only consider which migrations are in our list of
configured mappings. However, we have to consider the containers that
a node may have handled previously and is no longer responsible for
(this may happen if a container is added or removed).

The change will check all of the containers the node actually handled
and compare it to the existing status list, removing any containers that
were not processed.