Allow delete remote from local
New configuration flag for containers: allow_delete_remote_from_local

The purpose of this flag is to allow an archived object to be removed by
performing a delete on the local store. i.e. The user sees an archived
object through the shunt's merged namespace and deletes it. Rather than
getting 404, the shunt proceeds to delete the remote object.