Limit selection based on metadata
This introduces a new flag for mappings which allows the specification of
a filter to limit the objects moved based on metadata criteria.

The selection_criteria option allows robust definition including AND, OR, and
NOT blocks. The value is required to be a JSON object with a single key/value
pair. If the key is one defined below, the value must be a list or an object
as specified. Otherwise, the key/value pair is used as the match criteria.

Special keys:
    'AND' - the value must be a list of selection criteria that all must match
            for the selection criteria to match.
    'OR' - the value must be a list of selection criteria, any of which match
           will mean that the criteria matches.
    'NOT' - the value is a single selection criteria, reversing the sense of
            the match logic.

To specify "Only archive objects with 'x-object-meta-foo' set to 'bar'",
you could use:

    'selection_criteria': {'x-object-meta-foo': 'bar'}

To specify "Only archive objects with 'x-object-meta-foo' set to 'bar' unless
those objects have 'x-object-meta-other' set to 'baz', you could use:

    'selection_criteria': {'AND': [{'x-object-meta-foo': 'bar'},
                                   {'NOT': {'x-object-meta-other': 'baz'}}]}

Co-authored-by: Timur Alperovich <timuralp@swiftstack.com>