Migrator: handle corrupted status files.
Status files may be corrupted if the migrator is killed in the process
of writing then. The change makes two changes:
1. status files are now written to a temporary file and then renamed
2. if the status file is corrupted, the migrator moves it with a
   corrupted suffix and restarts

While the first change may be sufficient to prevent the situations of
the processing being killed in the middle of a write, we should still be
robust to something mutating the file as it is stored on disk.