Properly detect uploaded Swift SLOs in Swift.
When uploading Static Large Objects (SLO) to a remote Swift cluster, our
current check of whether an object has already been uploaded is
insufficient. Currently, we check if the SLO manifest ETags match, but
this is insufficient if the segments containers differ.

The patch changes our check to construct the expected ETag as a hash of
hashes and compare this ETag with the remote SLO ETag (not the manifest
ETag). This is similar to how we compare the MPU uploads in S3.