HEAD segments before uploading.
We should check if a segment has already been uploaded beforehand. This
allows us to restart SLO uploads on failures, handle SLOs that reference
the same segments, and more gracefully handle the case of multiple
daemons picking up the same SLO (in the worst case, it's the same
behavior, but if one manages to make significant progress, we will avoid
duplicate PUTs of segments).