Add the statsd plumbing for copied objects. (#190)
Adds the metric for objects copied by swift-s3-sync. There is follow-on
work where we want to distinguish PUT vs already existing objects, but
deferring that to a subsequent patch. We also want to similarly expose
the number of objects copied, but that is similarly deferred to a
follow-on patch.