Split S3Proxy into its own container.
Moves S3Proxy into a separate container that is built outside of the
main run_tests cycle. This makes the tests smaller and allows us to run
off SNAPSHOT versions of jclouds and S3Proxy. As a result, we pull in
the MPU ETag patch in the filesystem provider, which allows us to make
our tests more faithful.