Change the list_buckets API to use keyword args.
Currently, the list_buckets API requires delimiter, marker, and prefix
arguments. These are not supported in S3 bucket listings. The change
allows a provider to implement a richer API if it so chooses, but by
default exposes a narrow interface.

The named arguments also help us disambiguate the calls and I believe
are more clear.