Sync X-Delete-At header.
Added a new option to propagate the expiration header (Default: false)
When enabled, the `X-Delete-At` header is synced to remote location.
This option is only valid when syncing to a Swift target as S3 does not
support object expiration at the object level, only at the bucket level.
Furthermore, the new option is ignored if `remote_delete_after` is set.