Implement support for DLO in Swift sync/lifecycle.
Adds the ability to upload DLOs in sync/lifecycle mappings targeting
Swift as the remote blob store. A separate patch will implement
the DLO -> MPU conversion for S3 (including GCS).

The patch expands the public Swift provider API. Specifically, the
update_metadata() method now allows for an optional remote_metadata
parameter. This is necessary for DLO support, as we have to compare our
stored DLO header on the remote object (if any) and compare it with the
internal DLO ETag.