Add DLO support for the Swift provider for swift-s3-sync. (#253)
* Implement support for DLO in Swift sync/lifecycle.

Adds the ability to upload DLOs in sync/lifecycle mappings targeting
Swift as the remote blob store. A separate patch will implement
the DLO -> MPU conversion for S3 (including GCS).

The patch expands the public Swift provider API. Specifically, the
update_metadata() method now allows for an optional remote_metadata
parameter. This is necessary for DLO support, as we have to compare our
stored DLO header on the remote object (if any) and compare it with the
internal DLO ETag.

* Fixup for skipping nested SLOs.

Fixes the case of being stuck on nested SLOs. In general, we should
return a dictionary of statuses from the upload workers and we can then
use that for stats tracking.

* Add a test for Swift's remote account override.

It looks like we never actually hit an issue with URLs like
http:host/path, but it probably is better to fix this.