Implement support for minimum segment size. (#259)
Allows for conversion of objects with many small segments into ones
composed of larger segments. This applies both to DLO and SLO. In case
of DLO, we do not allow minimum segment size to be set without also
setting the convert_dlo option (converting DLOs to SLOs). This is
because we cannot match the DLO prefix and naming scheme otherwise.