Option to add an offset to remote X-Delete-At. (#264)
Optionally, add the specified expiration time to the object's set
expiration date. Applies only to Swift.

Alos, fixed an issue where X-Delete-At header is not propagated if the
object has already been uploaded without it. There is still an ongoing
problem with large objects' segments not being updated with the correct
expiration headers and the X-Delete-At header not being cleared when the
option is unset.