Use the listing timestamp if higher precision. (#273)
This addresses an issue where certain object stores (e.g. GCS) include
sub-second resolution in their listings, but it is not available when
doing HEAD on the object (through the S3 interoperability API -- higher
precision dates are available when using the JSON GCS API).

Until we consider moving to using the JSON API, we have to reconcile
this problem. The specific issue is that currently 1space migrations
will set the X-Timestamp value when retrieving an object to the object's
Last-Modified or X-Timestamp value. In the case of GCS, this will be
incorrect, as the listing time will be newer than the set X-Timestamp.

The patch fixes this problem by propagating the listing timestamps on
the record into the underlying upload methods, which can then compare it
to the headers retrieved via the GET/HEAD request.