Allow for empty "Body" or "Buckets" in SyncS3. (#274)
Boto may return an empty "Body" or "Buckets" keys when there is no
actual content. In case of a 0-sized object, this doesn't affect us, as
we will return an empty string. However, when returning results from an
account with no buckets, we should still return an empty list (as
opposed to an empty string).