Add statsd output on the migrator (#272)
This revamps the statsd collection code to enable stats collection in the migrator. It re-commonizes the work already done with the container sync code so they share some commonality. Future work will add additional statistics to be collected, but for now we collect things that we already had good code in place to easily gather: bytes transferred, objects copied, and objects scanned.