Add tests for migrations that use a prefix. (#281)
Adds a test that migrates only a subset of data in an S3 bucket using
the prefix setting.

As part of this change, we move to using unicode strings when calling
the providers and in the shunt. This was uncovered when the prefix
argument was expected to be a byte string in the provider's list_objects
method, but is a unicode string when reading the config.