Allow setting Storage Policy for sync/lifecycle.
The storage policy is only used when new containers are created (when
performing a '/*' migration or sync, for instance).

To test this change, a new storage policy is added to the container.
Similarly, the verification comman is expanded to allow for the new
option. The option is only valid for Swift endpoints.